requirements.txt
pytorch==1.0.1
numpy==1.18.1
torchvision
opencv